refactor(dsp): rename modules to reflect http
