--- data-protocols/dsp/dsp-core/build.gradle.kts
@@ -1,22 +1,0 @@
-/*
- *  Copyright (c) 2023 Fraunhofer Institute for Software and Systems Engineering
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation
- *
- */
-
-plugins {
-    `java-library`
-}
-
-dependencies {
-    api(project(":spi:common:http-spi"))
-    api(project(":data-protocols:dsp:dsp-spi"))
-}
