--- data-protocols/dsp/dsp-spi/src/main/java/org/eclipse/edc/protocol/dsp/spi/types/MessageProtocol.java
@@ -1,24 +1,0 @@
-/*
- *  Copyright (c) 2023 Fraunhofer Institute for Software and Systems Engineering
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation
- *
- */
-
-package org.eclipse.edc.protocol.dsp.spi.types;
-
-/**
- * Provides the name to be used as reference to the dataspace protocol in remote messages.
- */
-public class MessageProtocol {
-    
-    public static final String DATA_SPACE_PROTOCOL = "dataspace-protocol";
-    
-}
