+++ settings.gradle.kts
@@ -67,8 +67,8 @@ include(":core:data-plane-selector:data-plane-selector-core")
 
 // modules that provide implementations for data ingress/egress ------------------------------------
 include(":data-protocols:dsp:dsp-api-configuration")
-include(":data-protocols:dsp:dsp-core")
-include(":data-protocols:dsp:dsp-spi")
+include(":data-protocols:dsp:dsp-http-core")
+include(":data-protocols:dsp:dsp-http-spi")
 include(":data-protocols:dsp:dsp-transform")
 
 include(":data-protocols:ids:ids-api-configuration")
